+++ wgpu-hal/src/gles/command.rs
@@ -780,6 +780,10 @@ impl crate::CommandEncoder for super::CommandEncoder {
         self.rebind_sampler_states(dirty_textures, dirty_samplers);
     }
 
+    unsafe fn clear_bind_group(&mut self, _layout: &super::PipelineLayout, _index: u32) {
+        // TODO.
+    }
+
     unsafe fn set_push_constants(
         &mut self,
         _layout: &super::PipelineLayout,