+++ wgpu-hal/src/vulkan/command.rs
@@ -820,6 +820,9 @@ impl crate::CommandEncoder for super::CommandEncoder {
             )
         };
     }
+    unsafe fn clear_bind_group(&mut self, _layout: &super::PipelineLayout, _index: u32) {
+        // TODO.
+    }
     unsafe fn set_push_constants(
         &mut self,
         layout: &super::PipelineLayout,